backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Amine SGHIR <sghir.ma@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
django = "^5.0.7"
django-hosts = "^6.0"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
django-allauth = "^0.63.6"
dj-rest-auth = "^6.0.0"
django-import-export = "^4.1.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
django-post-office = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
